api-supabase/CDL_Jovem_Vila_Velha_API/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5
python-dotenv==1.0.0
httpx==0.28.1
pandas==2.3.0
python-multipart==0.0.20 